requirements.txt
sqlalchemy >= 1.0
six
typing; python_version < '3.5'
arrow
phonenumbers
python-slugify